feat: SEO 메타데이터, favicon 및 페이지별 title 추가 (#85)
- app/layout.tsx: title template, OG/Twitter 메타데이터, lang="ko" 설정
- 각 protected 페이지 metadata export 추가 (dashboard, repos, pulls, comments, notifications, stats, settings)
- pulls/[id]: generateMetadata로 동적 title 적용
- app/favicon.ico, icon.svg, apple-icon.svg: 로봇 테마 브랜드 아이콘으로 교체
- notifications/page.tsx, repositories/page.tsx: 서버 컴포넌트로 분리 (NotificationsClient, RepositoriesClient)
